+++ src/OpenTelemetry.AutoInstrumentation/Constants.cs
@@ -28,7 +28,7 @@ internal static class Constants
     {
         public const string AttributeServiceName = "service.name";
         public const string AttributeServiceInstanceId = "service.instance.id";
-        public const string AttributeServiceNamespaceName = "service.namespace.name";
+        public const string AttributeServiceNamespace = "service.namespace";
     }
 
     public static class ConfigurationValues